finished running on rcc

+++ submission/mv_iwas/prep_ukb_idp_gwas.sh
@@ -3,6 +3,7 @@
 # args3: tag
 # args4: outdir
 # args5: new position file
+# args6: download only
 
 url=$1
 fn=$2
@@ -12,9 +13,16 @@ newposition=$5
 
 cd $outdir
 
-wget $url -O $fn
-gunzip 0019.txt.gz
-zcat $fn | awk -F'\t' 'NR==1{print;next}; {$4=10**(-1*$4); print}' > $tag.tmp.txt
-paste $newposition $tag.tmp.txt > $tag.txt
-sed -e '1s/RSID/SNP/' -e '1s/PVAL/P/' -e '1s/CHROM/CHR/' $tag.txt > $tag.tmp2.txt
 
+if [[ ! -f $fn ]]
+then
+  wget $url -O $fn
+fi 
+
+if [[ -z $6 && ! -f $tag.tmp2.txt ]]
+then
+  # gunzip $fn
+  zcat $fn | awk -F'\t' 'NR==1{print;next}; {$4=10**(-1*$4); print}' > $tag.tmp.txt
+  paste $newposition $tag.tmp.txt > $tag.txt
+  sed -e '1s/RSID/SNP/' -e '1s/PVAL/P/' -e '1s/CHROM/CHR/' $tag.txt > $tag.tmp2.txt
+fi
